Apply check if file exists, remove file if it does

+++ project/SPT.Custom/Patches/EasyAssetsPatch.cs
@@ -97,7 +97,10 @@ namespace SPT.Custom.Patches
                     // only download when connected externally
                     if (await BundleManager.ShouldAcquire(bundleInfo))
                     {
-                        VFS.DeleteFile(BundleManager.GetBundleFilePath(bundleInfo));
+                        if (VFS.Exists(BundleManager.GetBundleFilePath(bundleInfo)))
+                        {
+                            VFS.DeleteFile(BundleManager.GetBundleFilePath(bundleInfo));
+                        }
 
                         await BundleManager.DownloadBundle(bundleInfo);
                     }
